% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{my_hello}
\alias{my_hello}
\title{hello}
\usage{
my_hello(name, x)
}
\arguments{
\item{name}{is your name}

\item{x}{is a number from 1 to 10, try it}
}
\value{
Returns greetings and something else
}
\description{
Salutations, I am just trying to make a package
}
\examples{
\dontrun {
  my_hello(name,x)
}

}
\author{
Me
}
