% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_subset_dt.R
\name{s2.subset.dt}
\alias{s2.subset.dt}
\title{s2.subset.dt}
\usage{
s2.subset.dt(string, data)
}
\arguments{
\item{string}{A character string to be mached to the \code{id} column. Examples: \code{ENSG} to select genes (if Ensembl ids are used), \code{chr} for ATAC.seq peaks.}

\item{data}{A \code{data.table} with an \code{id} column.}
}
\value{
A \code{data.table} containing only the rows where the 'id' column matches \code{string}.
}
\description{
Subset a data.table by matching a character with the "id" column.
}
\author{
Maiolino_Aurelio
}
