% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s5_GR_protein_coding.R
\name{s5.GR.protein.coding}
\alias{s5.GR.protein.coding}
\title{s5.GR.protein.coding}
\usage{
s5.GR.protein.coding(
  path = "/sharedFolder/Data/Homo_sapiens.GRCh38.114.gtf.gz"
)
}
\arguments{
\item{path}{Path to the GTF annotation file (default: \code{"/sharedFolder/Data/Homo_sapiens.GRCh38.114.gtf.gz"}).}
}
\value{
A \code{GRanges} object containing only rows where \code{type == "gene"} and \code{gene_biotype == "protein_coding"}.
}
\description{
Imports a GTF file and returns a \code{GRanges} object containing only protein-coding genes.
Standardizes the chromosome names to include "chr"
}
\author{
Maiolino_Aurelio
}
