% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s5_remap_atac.R
\name{s5.remap.atac}
\alias{s5.remap.atac}
\title{s5.remap.atac}
\usage{
s5.remap.atac(atac_data = gr_atac)
}
\arguments{
\item{atac_data}{A \code{GRanges} object containing ATAC-seq peaks. Defaults to \code{gr_atac}.}
}
\value{
A \code{GRanges} object with an additional column \code{overlapping_genes} listing overlapping gene IDs.
}
\description{
#' Remap ATAC Peaks to Overlapping Coding Genes

Finds overlaps between ATAC-seq peaks and coding genes, and annotates each peak
with the gene(s) it overlaps. If a peak overlaps multiple genes, the gene IDs are
concatenated in a comma-separated list.
}
\author{
Maiolino_Aurelio
}
