% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s6_finalize_exp_data.R
\name{s6.finalize.exp.data}
\alias{s6.finalize.exp.data}
\title{s6.finalize.exp.data}
\usage{
s6.finalize.exp.data(exp_data = gr_genes)
}
\arguments{
\item{exp_data}{A \code{GRanges} object containing gene expression data, with a \code{gene_id} column. Defaults to \code{gr_genes}.}
}
\value{
A \code{GRanges} object with an added \code{gene_symbol} column and filtered rows where gene symbol could not be assigned.
}
\description{
Finalize Expression Data with Gene Symbols

Annotates expression data by adding gene symbols based on gene IDs, using a reference annotation.
Rows without a matching gene symbol are removed from the returned dataset.
}
\author{
Maiolino_Aurelio
}
