% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_normalize_cmp.R
\name{s7.normalize.cmp}
\alias{s7.normalize.cmp}
\title{s7.normalize.cmp}
\usage{
s7.normalize.cmp(data)
}
\arguments{
\item{data}{A numeric vector of counts to be normalized and log-transformed.}
}
\value{
A numeric vector of normalized and log2-transformed values.
}
\description{
Normalize and Log-Transform Count Data

Normalizes numeric data to counts per million (CPM) and applies a log2 transformation.
This is commonly used for normalizing and transforming count data such as gene expression counts.
}
\author{
Maiolino_Aurelio
}
