% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_summary_chr_wise.R
\name{s7.summary.chr.wise}
\alias{s7.summary.chr.wise}
\title{s7.summary.chr.wise}
\usage{
s7.summary.chr.wise()
}
\value{
A \code{data.table} with columns \code{chr}, \code{N_unmerged_genes}, and \code{N_unmerged_peaks}, representing the chromosome, number of genes without ATAC peaks, and number of unmerged ATAC peaks, respectively.
}
\description{
Chromosome-wise Summary of Unmerged Peaks and Genes

Summarizes the count of unmerged ATAC-seq peaks and genes without ATAC signal for each chromosome.
}
\author{
Maiolino_Aurelio
}
