% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_summary_table_genes.R
\name{s7.summary.table.genes}
\alias{s7.summary.table.genes}
\title{s7.summary.table.genes}
\usage{
s7.summary.table.genes(exp_data = gr_genes, merged = merged_data)
}
\arguments{
\item{exp_data}{A \code{GRanges} object or data.table containing gene expression data, with columns \code{gene_id}, \code{gene_symbol}, and \code{gene_cpm}. Defaults to \code{gr_genes}.}

\item{merged}{A \code{data.table} containing merged gene expression and ATAC-seq data, with columns \code{gene_id}, \code{gene_symbol}, and \code{peak_cpm}. Defaults to \code{merged_data}.}
}
\value{
A \code{data.table} of genes without associated ATAC peaks, with columns: \code{gene_id}, \code{gene_symbol}, and \code{gene_cpm}.
}
\description{
Generate Summary Table of Genes Without ATAC Signal

Creates a summary table of genes that do not have overlapping ATAC-seq signal or have a peak CPM of zero.
The resulting table includes gene ID, gene symbol, and gene CPM for each gene without corresponding ATAC signal.
}
\author{
Maiolino_Aurelio
}
