% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somma.R
\name{somma}
\alias{somma}
\title{somma}
\usage{
somma(x, y)
}
\arguments{
\item{x}{is a numeric vector}

\item{y}{is a numeric vector}
}
\value{
the sum of x and y
}
\description{
a simple sum function
}
\examples{
\dontrun {
  sum(x,y)
}

}
\author{
Me
}
